backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.8.1
celery==5.4.0
redis==5.2.1
requests==2.32.3
pypdf==5.3.0
python-docx==1.1.2
openpyxl==3.1.5
pypdfium2==5.5.0
Pillow==11.1.0
pytesseract==0.3.13
python-multipart==0.0.20
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
pytest==8.3.4
httpx==0.28.1
langchain-core==0.3.34
langgraph==0.2.61
